requirements.txt
cassandra-driver==3.3.0
futures==3.0.5
six==1.10.0
